requirements.txt
customtkinter==5.2.2
opencv-python==4.9.0.80
mediapipe==0.10.11
ai-edge-litert
vosk==0.3.45
pyttsx3==2.90
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
Pillow==10.2.0
